Refactor scopes
