+++ apps/web/utils/gmail/scopes.ts
@@ -10,3 +10,16 @@ export const SCOPES = [
     ? ["https://www.googleapis.com/auth/contacts"]
     : []),
 ];
+
+export const CALENDAR_SCOPES = [
+  "https://www.googleapis.com/auth/calendar.events.readonly",
+  "https://www.googleapis.com/auth/calendar.events",
+  "https://www.googleapis.com/auth/calendar.readonly",
+  "https://www.googleapis.com/auth/calendar",
+  // Advanced calendar scopes for future use:
+  // "https://www.googleapis.com/auth/calendar.events.freebusy", // Read free/busy information for events
+  // "https://www.googleapis.com/auth/calendar.settings.readonly", // Read calendar settings (timezone, etc.)
+  // "https://www.googleapis.com/auth/calendar.settings", // Read and modify calendar settings
+  // "https://www.googleapis.com/auth/calendar.addons.execute", // Execute calendar add-ons
+  // "https://www.googleapis.com/auth/calendar.freebusy", // Read free/busy information for calendars
+];
